chore: Restructure events signal and enums
- Restructured event-related signals for better organization.
- Refactored and restructured enums for consistency and clarity.
- Added restriction to allow manipulation only for draft events.
